refactor(asset): move all loaders into asset/loader/, unprefix filenames
