+++ src/asset/loader/fbx-text.ts
