+++ src/asset/loader/fbx.ts
