+++ src/asset/loader/obj.ts
